css: move out toObject function
